Moved image metadata fetch to _test pkg to void cyclic imports.
